:hammer: Refactor: utils 함수 분리, interface 분리
