+++ src/lib/types/index.ts
@@ -26,7 +26,7 @@ export interface LoginBody {
   password: string;
 }
 
-// 회원가입
+// 회원가입 (API)
 export interface SignUpBody {
   email: string;
   password: string;
@@ -36,13 +36,54 @@ export interface SignUpBody {
   deptId: number;
 }
 
-//비밀번호 변경
-export interface editPasswordBody {
+// 회원가입 (Component)
+export interface SignUpForm {
+  email: string;
+  password: string;
+  pwCheck: string;
+  name: string;
+  hospital: string;
+  dept: string;
+  phone: string;
+}
+
+export interface Hospital {
+  hospitalName: string;
+  hospitalId: number;
+}
+
+export interface Department {
+  deptName: string;
+  deptId: number;
+}
+
+// 비밀번호 변경 (API)
+export interface EditPasswordBody {
   newPassword: string;
   oldPassword: string;
 }
 
-//마이페이지 수정
+// 비밀번호 변경 (Component)
+export interface EditPasswordForm {
+  oldPassword: string;
+  newPassword: string;
+  pwCheck: string;
+}
+
+// 개인정보 수정
+export interface ProfileBody {
+  name: string;
+  password: string;
+  deptName: string;
+  phone: string;
+  originImage: FileList;
+}
+
+export interface Password {
+  password: string;
+}
+
+// 마이페이지 수정
 export interface EditMyPageBody {
   name: string;
   deptId: number;
@@ -50,32 +91,37 @@ export interface EditMyPageBody {
   image: string | null | unknown;
 }
 
-//연차 등록
+// 연차 등록
 export interface CreateAnnualBody {
   startDate: Date;
   endDate: Date;
   reason: string;
 }
 
-//연차 내용 수정
+// 연차 내용 수정
 export interface EditAnnualBody {
   startDate: Date;
   endDate: Date;
   reason: string;
 }
 
-//당직 등록
+// 연차 신청 취소
+export interface CancelAnnualBody {
+  id: number;
+}
+
+// 당직 등록
 export interface CreateDutyBody {
   startDate: Date;
 }
 
-//당직 내용 수정
+// 당직 내용 수정
 export interface EditDutyBody {
   startDate: Date;
   updateDate: Date;
 }
 
-//캘린더 조회
+// 캘린더 조회
 export interface Schedule {
   category: string;
   deptName: string;
@@ -88,7 +134,7 @@ export interface Schedule {
   startDate: string;
 }
 
-//날짜별 휴가 조회
+// 날짜별 휴가 조회
 export interface AnnualData {
   deptName: string;
   id: number;
@@ -97,7 +143,7 @@ export interface AnnualData {
   username: string;
 }
 
-//날짜별 당직 조회
+// 날짜별 당직 조회
 export interface DutyData {
   deptName: string;
   email: string;
@@ -107,4 +153,60 @@ export interface DutyData {
   profileImageUrl: string;
   userId: number;
   username: string;
+}
+
+// 모달
+export interface ProfileProps {
+  $imgurl: null | string;
+}
+
+export interface DataBody {
+  id: number;
+  startDate: Date;
+  endDate: Date;
+  reason: string;
+}
+
+export interface ModalBtnProps {
+  handler: () => void;
+}
+
+// 사이드바
+export interface MenuItemProps {
+  to: string;
+  onClick?: () => void;
+  isactive?: string;
+}
+
+export interface SubMenuProps {
+  open?: boolean;
+}
+
+export interface ProgressProps {
+  $percent: number;
+}
+
+// 대시보드
+export interface DashBoardProps {
+  day: string;
+  week: string;
+  month: string;
+}
+
+// Decode
+export interface HospitalDecode {
+  hospital: string;
+  dept: DeptDecode;
+}
+
+export interface HospitalListDecode {
+  [key: number]: HospitalDecode;
+}
+
+export interface DeptDecode {
+  [key: number]: string;
+}
+
+export interface DeptNameDecode {
+  [key: number]: string;
 }
